feat: Bulk evaluation with list of flags
